Update magCIF.dic (#24)
Used comma-separated rather than space-separated vector strings in the transformation examples to match the usage in the symCIF dictionary.  Clarified that the tranformation-string format separates the point and translational components with a semicolon (all Pp type tags).  Clarified that the parent space-group to child transformation applies to the basic space group when the child is incommensurate.